+++ lib/providers/recording_box_visibility_provider.dart
@@ -1,0 +1,3 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+final recordingBoxVisibilityProvider = StateProvider<bool>((ref) => false);
